refactor: restructure project and fix error dommain

+++ app/Repositories/CustomerRepository.php
@@ -14,7 +14,7 @@ class CustomerRepository implements CustomerRepositoryInterface
         $this->model = $model;
     }
 
-    public function getAllPaginated($perPage = 10)
+    public function paginate($perPage = 10)
     {
         return $this->model->with('user')->latest()->paginate($perPage);
     }
@@ -24,12 +24,12 @@ class CustomerRepository implements CustomerRepositoryInterface
         return $this->model->with(['user', 'contracts'])->findOrFail($id);
     }
 
-    public function create(array $data)
+    public function create($data)
     {
         return $this->model->create($data);
     }
 
-    public function update($id, array $data)
+    public function update($id, $data)
     {
         $customer = $this->findById($id);
         $customer->update($data);
@@ -39,7 +39,22 @@ class CustomerRepository implements CustomerRepositoryInterface
     public function delete($id)
     {
         $customer = $this->findById($id);
-        $customer->user()->delete(); 
-        return $customer->delete();  
+        // Xóa user liên quan nếu là tài khoản portal
+        if ($customer->user) {
+            $customer->user()->delete(); 
+        }
+        return $customer->delete();
+    }
+
+    public function getSelectable()
+    {
+        return $this->model->select('id', 'company_name','tax_code','user_id')
+        ->with('user:id,full_name,email')
+        ->get();
+    }
+
+    public function findByUserId($userId)
+    {
+        return $this->model->where('user_id', $userId)->first();
     }
 }
